Move logger export to separate file to avoid errors
