+++ .env.example
@@ -24,5 +24,6 @@ SSL_CERT_PATH=
 SSL_KEY_PATH=
 PORT=
 
-# Log level, defaults to info if not declared
+# Log configuration, default values shown here
 LOG_LEVEL=info
+LOG_FILE=aggregator.log
