+++ src/AggregatorGateway.ts
@@ -21,7 +21,7 @@ import { AlphabillClient } from './consensus/alphabill/AlphabillClient.js';
 import { IAlphabillClient } from './consensus/alphabill/IAlphabillClient.js';
 import { LeaderElection } from './highAvailability/LeaderElection.js';
 import { LeadershipStorage } from './highAvailability/LeadershipStorage.js';
-import logger from './index.js';
+import logger from './Logger.js';
 import { RoundManager } from './RoundManager.js';
 import { ISmtStorage } from './smt/ISmtStorage.js';
 import { SubmitCommitmentStatus } from './SubmitCommitmentResponse.js';
