+++ src/highAvailability/LeaderElection.ts
@@ -1,7 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
 
 import { ILeadershipStorage } from './ILeadershipStorage.js';
-import logger from '../index.js';
+import logger from '../Logger.js';
 
 interface ILeaderElectionOptions {
   heartbeatInterval: number; // How often to send heartbeats
