+++ tests/TestContainers.ts
@@ -2,19 +2,21 @@ import { MongoDBContainer } from '@testcontainers/mongodb';
 import mongoose from 'mongoose';
 import { StartedTestContainer } from 'testcontainers';
 
-import logger from '../src/index.js';
+import logger from '../src/Logger.js';
 
 export async function startMongoDb(): Promise<StartedTestContainer> {
   const container = await new MongoDBContainer('mongo:7').start();
   const uri = container.getConnectionString();
-  logger.info(`Connecting to MongoDB, using connection URI: ${uri}.`);
-  await mongoose.connect(uri, { serverSelectionTimeoutMS: 5000, directConnection: true });
-  logger.info('Connected successfully.');
+  logger.info(`Connecting to MongoDB URI %s.`, uri);
+  await mongoose
+    .connect(uri, { serverSelectionTimeoutMS: 5000, directConnection: true })
+    .then(() => logger.info('Connected successfully.'))
+    .catch((err) => logger.error('Failed to connect to MongoDB: ', err));
   return container;
 }
 
 export async function stopMongoDb(container: StartedTestContainer): Promise<void> {
-  logger.info('\nStopping MongoDB container...');
+  logger.info('Stopping MongoDB container...');
   await container.stop({ timeout: 10 });
   logger.info('Test completed.');
 }
