+++ tests/consensus/alphabill/MockAlphabillClient.ts
@@ -24,7 +24,7 @@ import { DataHash } from '@unicitylabs/commons/lib/hash/DataHash.js';
 
 import { IAlphabillClient } from '../../../src/consensus/alphabill/IAlphabillClient.js';
 import { SubmitHashResponse } from '../../../src/consensus/alphabill/SubmitHashResponse.js';
-import logger from '../../../src/index.js';
+import logger from '../../../src/Logger.js';
 
 class MockSigningService implements ISigningService {
   public publicKey = new Uint8Array([1, 2, 3, 4]);
